modify SelectModules for ot-2 only
